requirements.txt
numpy==1.21.1
Pillow==8.4.0
scipy==1.3.1
stl==0.0.3
